Back_End/requirements.txt
fastapi
uvicorn[standard]
tensorflow
#tflite-runtime
numpy
pandas
Pillow
python-multipart
opencv-python
sqlalchemy
cryptography #?